working on driver endpoints

+++ controllers/driverController.js
@@ -20,16 +20,25 @@ exports.createDriver = async (req, res) => {
         return res.status(400).send("please all fields are required");
     }
 
+
     const ambulance = await Ambulance.findById(ambulanceId);
     
     if (!ambulance) {
       return res.status(404).json({ error: 'Ambulance not found' });
     }
 
+    const findDriverExists= await Driver.findOne({license, ambulanceId});
+
+    if(findDriverExists){
+      return res.status(401).send("driver  already exists");
+    }
+
     const newDriver = new Driver({ name, license, ambulance: ambulance._id });
-    await newDriver.save();
-    
-    return res.status(201).json({message:"driver has been created successfully",newDriver}); 
+   if( await newDriver.save()){
+     return res.status(201).json({message:"driver has been created successfully",newDriver}); 
+   }
+   return res.status(401).send("failed to create driver");
+
   } catch (err) {
     logger.error('Error creating driver:', err);
     res.status(500).json({ error: 'Internal server error' });
@@ -40,19 +49,30 @@ exports.updateDriver = async (req, res) => {
   try {
     const { id } = req.params;
     const { name, license, ambulanceId } = req.body;
+
+    if(!name||!license||!ambulanceId){
+      return res.status(400).send("please all fields are required");
+  }
+
     const ambulance = await Ambulance.findById(ambulanceId);
     if (!ambulance) {
       return res.status(404).json({ error: 'Ambulance not found' });
+    }
+
+  const findDriverExists= await Driver.findById(id);
+    if(!findDriverExists){
+      return res.status(401).send("driver not found");
     }
     const updatedDriver = await Driver.findByIdAndUpdate(
       id,
       { name, license, ambulance: ambulance._id },
       { new: true }
     );
+
     if (!updatedDriver) {
-      return res.status(404).json({ error: 'Driver not found' });
+      return res.status(404).json({ error: ' failed to update the driver' });
     }
-    res.json(updatedDriver);
+    res.json({message:"driver updated",updatedDriver});
   } catch (err) {
     logger.error('Error updating driver:', err);
     res.status(500).json({ error: 'Internal server error' });
@@ -62,9 +82,15 @@ exports.updateDriver = async (req, res) => {
 exports.deleteDriver = async (req, res) => {
   try {
     const { id } = req.params;
+
+    const findDriverExists= await Driver.findOne(id);
+    if(!findDriverExists){
+      return res.status(401).send("driver not found");
+    }
+
     const deletedDriver = await Driver.findByIdAndDelete(id);
     if (!deletedDriver) {
-      return res.status(404).json({ error: 'Driver not found' });
+      return res.status(404).json({ error: 'failed to delete the driver' });
     }
     res.json({ message: 'Driver deleted' });
   } catch (err) {
